requirements.txt
camelot-py[cv]
pandas
pdfplumber
opencv-python
numpy
pillow
pymupdf